requirements_web.txt
# Web application dependencies
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-crispy-forms==2.4
django-environ==0.12.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.6.0
pillow==11.2.1

# Social Authentication
social-auth-app-django==5.5.1
social-auth-core==4.7.0

# Explicitly exclude TensorFlow and related packages
# These cause memory issues on Render